Introduce composite module, to compose storages easily.
* Changes:

- Introduce composite module, to compose storages into one.
- Introduce AllocatorBuilder, a Builder for all allocator storages.
- Use composite to reimplement small::SingleElement.
- Use composite to introduce small::SingleRange.

* Motivation:

The ability to quickly spin up a composite storage makes it much easier.

* Bonus:

Clean up the useless feature, there's no need for the allocator to
default to Global in these low-level modules, and since this was the
only non-core dependency (outside of tests), removing it allows simply
dropping the "alloc" feature, greatly simplifying things.