Introduce ElementStorage::allocate, to allow allocation without value
* Change:

- Introduce ElementStorage::allocate.
- Rename acquire/release to allocate/deallocate.
- Implement RawBox::try_in, to switch to another storage.

* Motivation:

Moving an Unsized value to another storage cannot be done with
ElementStorage::create, so just like destroy was supplemented by
deallocate, supplement create by allocate.

It is notable that create and destroy really are just methods of
convenience, now. Sugar is sweet.

As for allocate/deallocate, it matches better to the Allocator API, and
since it's essentially a thin wrapper over it, this seems like a good
thing.